Override start and stop streaming events
Provided that the documentation is still up-to-date, subscribing to the
channel:1:update event will fire chat:1:StartStreaming and
chat:1:StopStreaming events. Since these events are entirely different
from the subscription sent, this was not compatible with the library.

This commit overrides it so that listening for chat:1:StartStreaming
will silently subscribe to the “update” event instead.

This might have changed from Liveloading to Constellation though (since
I’m basing this commit off liveloading’s docs)